+++ src/components/home/KoreaMovie.tsx
@@ -1,23 +1,23 @@
 'use client';
 
 /**
- * Preview 섹션 컴포넌트
- * 인기 영화를 원형 이미지로 미리보기 형태로 표시합니다.
+ * KoreaMovie 섹션 컴포넌트
+ * 한국 영화를 표시합니다.
  */
 
 import { useEffect, useState } from 'react';
 
-import { getMoviePopular } from '@/apis/home';
 import MovieSwiper from '@/components/home/MovieSwiper';
+import { getKoreaMovie } from '@/lib/api';
 import { Movie } from '@/types/tmdb';
 
-const Preview = () => {
+const KoreaMovie = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     const fetchMovies = async () => {
-      const response = await getMoviePopular();
+      const response = await getKoreaMovie();
       setMovies(response.results);
       setLoading(false);
     };
@@ -25,11 +25,9 @@ const Preview = () => {
     fetchMovies();
   }, []);
 
-  if (loading) {
-    return null;
-  }
+  if (loading) return null;
 
-  return <MovieSwiper title="Previews" items={movies} itemWidth="102px" itemHeight="102px" shape="circle" />;
+  return <MovieSwiper title="Korea Movies" items={movies} itemWidth="103px" itemHeight="161px" />;
 };
 
-export default Preview;
+export default KoreaMovie;
