+++ src/components/home/Preview.tsx
@@ -1,23 +1,23 @@
 'use client';
 
 /**
- * TopRatedMovies 섹션 컴포넌트
- * 높은 평점의 영화를 표시합니다.
+ * Preview 섹션 컴포넌트
+ * 인기 영화를 원형 이미지로 미리보기 형태로 표시합니다.
  */
 
 import { useEffect, useState } from 'react';
 
-import { getTopRatedMovies } from '@/apis/home';
 import MovieSwiper from '@/components/home/MovieSwiper';
+import { getMoviePopular } from '@/lib/api';
 import { Movie } from '@/types/tmdb';
 
-const TopRatedMovies = () => {
+const Preview = () => {
   const [movies, setMovies] = useState<Movie[]>([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     const fetchMovies = async () => {
-      const response = await getTopRatedMovies();
+      const response = await getMoviePopular();
       setMovies(response.results);
       setLoading(false);
     };
@@ -25,11 +25,9 @@ const TopRatedMovies = () => {
     fetchMovies();
   }, []);
 
-  if (loading) {
-    return null;
-  }
+  if (loading) return null;
 
-  return <MovieSwiper title="Top Rated Movies" items={movies} itemWidth="103px" itemHeight="161px" />;
+  return <MovieSwiper title="Previews" items={movies} itemWidth="102px" itemHeight="102px" shape="circle" />;
 };
 
-export default TopRatedMovies;
+export default Preview;
