+++ src/lib/api/index.ts
@@ -1,0 +1,4 @@
+export { axiosInstance } from '@/lib/api/axiosInstance';
+export * from '@/lib/api/tmdb/detail';
+export * from '@/lib/api/tmdb/home';
+export * from '@/lib/api/tmdb/search';
